Don't create a Mollie customer when the gateway is inactive
A client had invalid credentials in the Gateway settings which resulted in a Mollie::RequestError. Let's only create a Mollie customer when the gateway is active.

Using an ensure_mollie_customer after_save should also create Mollie customers for users that have been created in the time that is was inactive